Attempt to move to sbt
I've tried to remove all of the java dependencies so we can build this
for scalajs/scala native.

Scala native might be close but I don't want to worry about cross
platform builds.

Scala js won't work because none of the java.nio.file classes are
available in ScalaJS.

We may do this in future but for now I'm abandoning it.