DR2-1969 assisted file path substitution
Fixes https://github.com/digital-preservation/csv-validator/issues/507

If you select a csv and a schema file and click settings -> Add path substitution
then it will search the file for a column called identifier. If it's
there, it will choose the shortest value in that column and populate the
path from field.

If you don't select a csv and schema file, it can't parse the csv to
find the column so it will show a blank text box.

The to field is a free text field but there's also a button for
selecting a folder. The text in the dialog says which folder to search
for. Once a folder is selected, it will url encode it, add a trailing slash
if it is missing and put it in the text box.

You can also paste a path in and click the file chooser and it will start from there,
otherwise, it will start from your home directory.