Account for either user modifying 'from' field or 'identifier' not being found
Previously, if the identifier was found and 'From' field populated with the path, if the 'From' field was modified afterwards,
then the button text would still have the original path in it. Also. if identifier wasn't found but a user added in the 'From', manually
afterwards, then the button would not have that 'From' path